+++ Antonieta.Backend/dtos/update_ration_stock_dto.py
@@ -8,4 +8,4 @@ class update_ration_stock_dto(BaseModel):
     name: str
     description: str
     unit: str
-    stock: int
+    stock: float
